Refactor (SampleApp): Extract MessageAdapter into it's own class - makes the SampleApp clearer in terms of what's needed to integrate FlexInput
